improve implementation and tests, adds by default shadow-root element to dom
