+++ __test__/compilers/tag-file.test.js
@@ -1,25 +1,17 @@
-var compiler = require('../../compilers/tag-file')
+const compiler = require('../../compilers/tag-file')
+const utils = require('../utils')
+
 test('tag-file', function () {
   var content = `<tag-name>
-  <script>
-    var constructorCode = true
-  </script>
-  <h1>html test</h1>
-</tag-name>
-`
-
-  var expectedCompiledCode = `/** @jsx createElement */ module.exports = require('organic-oval').define({
-  tagName: "tag-name",
-  tagLine: "",
-  script: function () {
-    var constructorCode = true
-  },
-  template: function (createElement) {
-    return <h1>html test</h1>
-  }
-})
-`
-
+    <script>
+      var constructorCode = true
+    </script>
+    <h1>html test</h1>
+  </tag-name>`
   var compiled = compiler.compile(content)
-  expect(compiled.trim().replace(/ /g, '').replace(/\n/g, '')).toEqual(expectedCompiledCode.trim().replace(/ /g, '').replace(/\n/g, ''))
+  utils.expectTagFile(compiled, {
+    tagName: 'tag-name',
+    script: 'var constructorCode = true',
+    template: `<h1>html test</h1>`
+  })
 })
