+++ compilers/tag-file.js
@@ -1,4 +1,4 @@
-var extractScript = function (lines) {
+const extractScript = function (lines) {
   var buffer = []
   var scriptFound = false
   for (var i = 0; i < lines.length; i++) {
@@ -24,7 +24,7 @@ var extractScript = function (lines) {
   return buffer.join('\n')
 }
 
-var extractTagInfo = function (lines) {
+const extractTagInfo = function (lines) {
   var firstLine = lines[0]
   var tagName = firstLine.split(' ')[0].replace('<', '').replace('>', '')
   var tagLine = firstLine.replace('<' + tagName, '').replace('>', '')
@@ -33,39 +33,168 @@ var extractTagInfo = function (lines) {
   return {tagName, tagLine}
 }
 
+const parseLoops = function (lines) {
+  for (var i = 0; i < lines.length; i++) {
+    if (lines[i].indexOf('<each ') !== -1) {
+      let openingTag = lines[i].trim().slice('<each'.length, -1)
+      let params = openingTag.split(' in ')[0].trim()
+      let model = openingTag.split(' in ')[1].trim().slice('{'.length, -1) // get the model, removing ${ and }
+      lines[i] = `{ ${model}.map((${params}) => `
+    }
+    if (lines[i].indexOf('</each>') !== -1) {
+      lines[i] = lines[i].replace('</each>', ')}')
+    }
+  }
+}
+
+const extractStatement = function (line) {
+  // line example: 'class="something" if={statement} id="1"'
+  // output: 'statement'
+  var ifMark = 'if={'
+  // find index of if mark
+  var statementStart = line.indexOf(ifMark)
+  // find index of end statement
+  var statementEnd = -1
+  var bcount = 1
+  for (var i = statementStart + ifMark.length; i < line.length; i++) {
+    if (line.charAt(i) === '{') bcount += 1
+    if (line.charAt(i) === '}') bcount -= 1
+    if (bcount === 0) {
+      statementEnd = i
+      break
+    }
+  }
+  // return statement substring without ifMark and closing bracket
+  var statement = line.substring(statementStart + ifMark.length, statementEnd)
+  return statement
+}
+const extractAttributes = function (condition, line) {
+  return line.replace('if={' + condition + '}', '')
+}
+const parseIfs = function (lines) {
+  // variable holding amount of `if` statements encountered during parsing
+  // used to construct `oid` value
+  let ifcount = -1
+
+  // iterate over all lines searching for if statements
+  for (var i = 0; i < lines.length; i++) {
+    if (lines[i].indexOf('if=') !== -1) {
+      ifcount += 1
+      lines[i] = lines[i].trim()
+
+      // test transform multiline if statement such as
+      // <p if=${}
+      //    class="">...
+      // into singleline <p if=${} class="">...
+      var closingArrowIndex = lines[i].lastIndexOf('>')
+      if (closingArrowIndex !== lines[i].length - 1) {
+        var lineBuffer = []
+        for (var j = i + 1; j < lines.length; j++) {
+          if (lines[j].indexOf('>') === -1) {
+            lineBuffer.push(lines[j].trim())
+            lines.splice(j, 1)
+            j -= 1
+          } else {
+            lineBuffer.push(lines[j].trim())
+            lines.splice(j, 1)
+            j -= 1
+            break
+          }
+        }
+        lines[i] += ' ' + lineBuffer.join('')
+      }
+
+      // get the opening tag, remove '<' and '>'
+      var line = lines[i].trim().slice(1, -1)
+      // split the line, the first element is the node name, the others are attributes
+      // example: span class="something" if={condition} id="1" -> ['span', 'class="something"', 'if={condition}', 'id="1"']
+      var parsedLine = line.split(' ')
+
+      // get the node name
+      var nodeName = parsedLine.shift()
+      parsedLine = parsedLine.join(' ')
+      var statement = extractStatement(parsedLine)
+      var attributes = extractAttributes(statement, parsedLine)
+
+      let openBr = '{'
+      let closeBr = '}'
+
+      if (lines[i].indexOf('/>') === -1 && lines[i].indexOf('</' + nodeName + '>') === -1) {
+        // buffer lines between start and end closing tag when not on the same line
+        var buffer = []
+        var innerNodeNameCount = 0
+
+        for (var k = i + 1; k < lines.length; k++) {
+          // found a line with the same node name as the opening tag
+          if (lines[k].indexOf('<' + nodeName) !== -1) {
+            innerNodeNameCount += 1
+          }
+          // found a line with a name of the closing tag
+          if (lines[k].indexOf('</' + nodeName + '>') !== -1) {
+            if (innerNodeNameCount > 0) {
+              // the line has closing tag of a inner childrens
+              innerNodeNameCount -= 1
+            } else {
+              // the line is the closing tag of the statement
+              lines.splice(k, 1)
+              break
+            }
+          }
+          // buffer lines and remove them from original source
+          buffer.push(lines[k].trim())
+          lines.splice(k, 1)
+          k -= 1
+        }
+        // re-construct attributes with oid marker included
+        if (attributes.indexOf('key=') === -1) {
+          attributes = 'key={this.oid + "-if' + ifcount + '"} ' + attributes
+        }
+        // rewrite opening tag with if statement
+        lines[i] = openBr + statement + ' ? <' + nodeName + ' ' + attributes + '>'
+        // insert any buffered content at next line of `i + k`
+        for (let k = 0; k < buffer.length; k++) {
+          lines.splice(i + k + 1, 0, buffer[k])
+        }
+        // insert closing tag
+        lines.splice(i + buffer.length + 1, 0, '</' + nodeName + '> : null' + closeBr)
+      } else {
+        // single line if statement such as `<p if=${statement}>text</p>``
+
+        // get clean tag
+        var lineWithoutStatement = lines[i].replace('if={' + statement + '}', '')
+
+        // re-construct attributes with oid marker
+        let parts = lineWithoutStatement.trim().split(' ')
+        if (lineWithoutStatement.indexOf('key') === -1) {
+          parts.splice(1, 0, 'key={this.oid + "-if' + ifcount + '"}')
+        }
+        lineWithoutStatement = parts.join(' ')
+
+        // rewrite the line
+        lines[i] = openBr + statement + ' ? ' + lineWithoutStatement + ' : null' + closeBr
+      }
+    }
+  }
+}
+
 module.exports.compile = function (content) {
   var lines = content.trim().split('\n')
   var scriptContent = extractScript(lines)
   var tagInfo = extractTagInfo(lines)
+  parseIfs(lines)
+  parseLoops(lines)
   var htmlContent = lines.join('\n')
-
-  // @WORKAROUND
-  /** having this case happens when dealing with component
-  <my-comp>
-    <div if={statement}>...
-  </my-comp>
-
-  ie the if happens to be on the first element, therefore its
-  invalid js code as such
-  template: function () {
-    return {statement ? <div>... : null}
-  }
-
-  so we are doing a workaround to fix up after the `tag-control-statements`
-  */
-  if (htmlContent.indexOf('{') === 0) {
-    htmlContent = htmlContent.slice(1, -1)
-  }
-
   var result = `
   /** @jsx createElement */
 
   module.exports = require('organic-oval').define({
     tagName: "${tagInfo.tagName}",
-    tagLine: "${tagInfo.tagLine}",
-    script: function () { ${scriptContent.trim()} },
-    template: function (createElement) {
-      return ${htmlContent.trim()}
+    script: function () {
+      let tag = this
+      ${scriptContent.trim()}
+      this.template = function (createElement) {
+        return <shadow-root>${htmlContent.trim()}</shadow-root>
+      }
     }
   })
 `
