--- webpack/oval-control-statements-loader.js
@@ -1,15 +1,0 @@
-var TagCompiler = require('../compilers/tag-control-statements')
-
-module.exports = function (source) {
-  var content = source
-  if (this.cacheable) this.cacheable()
-  try {
-    return TagCompiler.compile(content)
-  } catch (e) {
-    if (e instanceof Error) {
-      throw e
-    } else {
-      throw new Error(e)
-    }
-  }
-}
